Conflicts:
	src/main/java/logic/ShopService.java
	src/main/java/logic/User.java
	src/main/webapp/WEB-INF/view/tutor/register-class.jsp

+++ src/main/java/logic/Course.java
@@ -19,8 +19,8 @@ public class Course {
 	private Date startdate;
 	private Date enddate;
 	private int reviewnum;
-	private int star;
-	
+	private double star;
+	 
 	public int getApplyno() {
 		return applyno;
 	}
@@ -111,10 +111,10 @@ public class Course {
 	public void setreviewnum(int reviewnum) {
 		this.reviewnum = reviewnum;
 	}
-	public int getStar() {
+	public double getStar() {
 		return star;
 	}
-	public void setStar(int star) {
+	public void setStar(double star) {
 		this.star = star;
 	}
 	@Override
